Added command to kick off Jenkins build
* Moved some Jenkins related API calls to a separate class (first pass)

Resolves #11
