+++ src/gluon/project/ProjectEnvironmentsRequested.ts
@@ -19,7 +19,7 @@ import {QMConfig} from "../../config/QMConfig";
 import {SimpleOption} from "../../openshift/base/options/SimpleOption";
 import {OCClient} from "../../openshift/OCClient";
 import {OCCommon} from "../../openshift/OCCommon";
-import {LinkExistingApplication} from "../packages/Applications";
+import {LinkExistingApplication} from "../packages/CreateApplication";
 
 @EventHandler("Receive ProjectEnvironmentsRequestedEvent events", `
 subscription ProjectEnvironmentsRequestedEvent {
